requirements.txt
python-dotenv
beautifulsoup4
praw
requests
google-auth
google-api-python-client